backend/requirements.txt
absl-py==2.4.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
aistudio-sdk==0.3.8
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.12.1
argbind==0.3.9
asgiref==3.11.0
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.1.0
Authlib==1.6.8
av==16.1.0
bce-python-sdk==0.9.57
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.2
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
contractions==0.1.73
crcmod==1.7
cryptography==46.0.5
ctranslate2==4.6.3
cycler==0.12.1
Cython==3.2.3
dataclasses-json==0.6.7
datasets==3.6.0
decorator==5.2.1
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
Django==5.2.8
dlib==20.0.0
dnspython==2.8.0
docstring_parser==0.17.0
ecdsa==0.19.1
editdistance==0.8.1
einops==0.8.2
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.133.1
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
faster-whisper==1.2.1
ffmpy==1.0.0
filelock==3.20.0
fire==0.7.1
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.3.0
funasr==1.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
gotrue==2.12.4
gradio==5.50.0
gradio_client==1.14.0
groovy==0.1.2
groq==0.36.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.27.2
httpx-sse==0.4.3
huggingface-hub==0.35.3
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
imgaug==0.4.0
inflect==7.5.0
jaconv==0.4.1
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaldifst==1.7.17
kaldiio==2.18.1
kiwisolver==1.4.9
langchain==1.0.7
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.5
langchain-huggingface==1.0.1
langchain-ollama==1.0.0
langchain-openai==1.0.3
langchain-pinecone==0.2.13
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
lmdb==1.7.5
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.32
modelscope==1.33.0
more-itertools==10.8.0
motor==3.7.1
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.15.0
networkx==3.5
nodeenv==1.10.0
noisereduce==3.0.3
numba==0.62.1
numpy==1.26.4
ollama==0.4.5
omegaconf==2.3.0
onnxruntime==1.23.2
openai==2.8.0
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.11.4
ormsgpack==1.12.0
oss2==2.19.1
outcome==1.3.0.post0
packaging==26.0
paddleocr==2.8.1
paddlepaddle==3.2.2
pandas==2.3.3
pandoc==2.4
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==11.0.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.5.0
plotly==6.5.2
plumbum==1.10.0
ply==3.11
pooch==1.8.2
postgrest==0.16.11
prettytable==3.17.0
prisma==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
py-cpuinfo==9.0.0
pyahocorasick==2.3.0
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
pyfiglet==1.0.4
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.15.4
PyMsgBox==2.0.1
pynndescent==0.6.0
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyparsing==3.2.5
pypdfium2==5.2.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytorch-wpe==0.0.1
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
realtime==1.0.6
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
rubicon-objc==0.5.3
ruff==0.14.14
safehttpx==0.1.7
safetensors==0.6.2
scikit-image==0.26.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
selenium==4.35.0
semantic-version==2.10.0
sentence-transformers==5.1.2
sentencepiece==0.2.1
sentry-sdk==2.44.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
sightengine==1.4.0
simplejson==3.20.2
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
soundfile==0.12.1
soupsieve==2.8.1
soxr==1.0.0
spaces==0.47.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
srt==3.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.52.1
storage3==0.7.7
streamlit==1.53.0
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.7
sympy==1.14.0
tenacity==9.1.2
tensorboardX==2.6.4
termcolor==3.3.0
textsearch==0.0.24
threadpoolctl==3.6.0
tifffile==2025.12.20
tiktoken==0.12.0
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.13.3
torch==2.10.0
torch-complex==0.4.4
torchaudio==2.10.0
torchcodec==0.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.1
transformers==4.36.2
tree-sitter-javascript==0.25.0
tree-sitter-typescript==0.23.2
trio==0.30.0
trio-websocket==0.12.2
typeguard==4.4.4
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
umap-learn==0.5.11
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.41.0
uvloop==0.22.1
vosk==0.3.44
voxcpm==1.5.0
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.8.0
websockets==12.0
wetext==0.1.2
wheel==0.45.1
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
